Cleanup and prepare main.py for Beacon node
There are several routes we can take to add
maintain an entirely different node type
in the same code base.

One pragmatic solution is to maintain multiple
main files and expose multiple top level commands

E.g. There would not only be `trinity` but also
`trinity2` (name to be discussed).

While this comes with the slight disadvantage of
occuping an entirely different command, it gives
us the most freedom for experimentation.

This change prepares our code base for this by
moving everything out of main.py that can be
re-used.
